#16102: Automatic spaces between CJK and Latin is inserted at the end of line
To add spaces between a CJK and Latin letter according to the position
in a line, this commit moves part of code adjusting those spaces to
pageitem_textframe.cpp. Now the apace is inserted before a ltter, in
a similar way to tracking.

It also introduces ScLayout_CJKLatinSpace, representing places
where that spaces are insterted.